Add OA Tools menu, separate dues upload from settings
